python/mvc-mvp-mvvm/mvc/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7